# Conflicts:
#	.gitignore
#	.travis.yml
#	README.md
#	baystation12.dme
#	code/__defines/_planes+layers.dm
#	code/__defines/atmos.dm
#	code/__defines/chemistry.dm
#	code/__defines/misc.dm
#	code/__defines/mobs.dm
#	code/__defines/subsystems.dm
#	code/_onclick/hud/_defines.dm
#	code/_onclick/hud/human.dm
#	code/datums/ai_law_sets.dm
#	code/datums/uplink/ammunition.dm
#	code/datums/uplink/devices and tools.dm
#	code/datums/uplink/grenades.dm
#	code/datums/uplink/hardsuit_modules.dm
#	code/datums/uplink/highly_visible_and_dangerous_weapons.dm
#	code/datums/uplink/stealth_and_camouflage_items.dm
#	code/game/antagonist/alien/xenomorph.dm
#	code/game/antagonist/outsider/mercenary.dm
#	code/game/antagonist/outsider/ninja.dm
#	code/game/antagonist/outsider/raider.dm
#	code/game/gamemodes/game_mode.dm
#	code/game/gamemodes/meteor/meteor.dm
#	code/game/gamemodes/mixed/bughunt.dm
#	code/game/machinery/Sleeper.dm
#	code/game/machinery/alarm.dm
#	code/game/machinery/autolathe.dm
#	code/game/machinery/bioprinter.dm
#	code/game/machinery/buttons.dm
#	code/game/machinery/constructable_frame.dm
#	code/game/machinery/cryopod.dm
#	code/game/machinery/doors/blast_door.dm
#	code/game/machinery/doors/door.dm
#	code/game/machinery/floor_light.dm
#	code/game/machinery/holosign.dm
#	code/game/machinery/machinery.dm
#	code/game/machinery/nuclear_bomb.dm
#	code/game/machinery/pipe/construction.dm
#	code/game/machinery/recharger.dm
#	code/game/machinery/robotics_fabricator.dm
#	code/game/machinery/status_light.dm
#	code/game/machinery/telecomms/logbrowser.dm
#	code/game/machinery/teleporter.dm
#	code/game/machinery/vending.dm
#	code/game/mecha/equipment/tools/tools.dm
#	code/game/mecha/mecha.dm
#	code/game/objects/effects/decals/cleanable.dm
#	code/game/objects/items/paintkit.dm
#	code/game/objects/items/stacks/tiles/tile_types.dm
#	code/game/objects/items/weapons/candle.dm
#	code/game/objects/items/weapons/cards_ids.dm
#	code/game/objects/items/weapons/gift_wrappaper.dm
#	code/game/objects/items/weapons/soap.dm
#	code/game/objects/items/weapons/tanks/tanks.dm
#	code/game/objects/structures.dm
#	code/game/objects/structures/stool_bed_chair_nest/chairs.dm
#	code/game/turfs/simulated/floor.dm
#	code/modules/admin/admin.dm
#	code/modules/admin/admin_verbs.dm
#	code/modules/admin/create_object.dm
#	code/modules/admin/topic.dm
#	code/modules/client/preference_setup/loadout/lists/misc.dm
#	code/modules/client/preference_setup/loadout/lists/tactical.dm
#	code/modules/clothing/chameleon.dm
#	code/modules/clothing/clothing.dm
#	code/modules/clothing/spacesuits/spacesuits.dm
#	code/modules/ext_scripts/irc.dm
#	code/modules/library/lib_items.dm
#	code/modules/library/manuals/engineering.dm
#	code/modules/library/manuals/medical.dm
#	code/modules/library/manuals/nanotrasen.dm
#	code/modules/materials/definitions/materials_other.dm
#	code/modules/mob/hear_say.dm
#	code/modules/mob/language/alien/antag.dm
#	code/modules/mob/language/alien/monkey.dm
#	code/modules/mob/language/generic.dm
#	code/modules/mob/living/carbon/alien/alien.dm
#	code/modules/mob/living/carbon/alien/larva/larva.dm
#	code/modules/mob/living/carbon/alien/larva/life.dm
#	code/modules/mob/living/carbon/carbon.dm
#	code/modules/mob/living/carbon/human/human_powers.dm
#	code/modules/mob/living/silicon/ai/ai.dm
#	code/modules/mob/living/silicon/pai/pai.dm
#	code/modules/mob/living/simple_animal/borer/borer.dm
#	code/modules/mob/living/simple_animal/borer/borer_powers.dm
#	code/modules/mob/living/simple_animal/friendly/lizard.dm
#	code/modules/mob/living/simple_animal/hostile/hivebot.dm
#	code/modules/mob/mob_defines.dm
#	code/modules/mob/mob_helpers.dm
#	code/modules/mob/mob_movement.dm
#	code/modules/modular_computers/computers/subtypes/preset_laptop.dm
#	code/modules/modular_computers/computers/subtypes/preset_pda.dm
#	code/modules/modular_computers/computers/subtypes/preset_tablet.dm
#	code/modules/modular_computers/file_system/computer_file.dm
#	code/modules/modular_computers/file_system/programs/command/card.dm
#	code/modules/modular_computers/file_system/programs/generic/docks.dm
#	code/modules/multiz/turf.dm
#	code/modules/organs/external/species/xenophage.dm
#	code/modules/organs/internal/appendix.dm
#	code/modules/organs/internal/species/xenophage.dm
#	code/modules/organs/robolimbs.dm
#	code/modules/paperwork/faxmachine.dm
#	code/modules/paperwork/photocopier.dm
#	code/modules/power/apc.dm
#	code/modules/power/breaker_box.dm
#	code/modules/power/port_gen.dm
#	code/modules/reagents/Chemistry-Machinery.dm
#	code/modules/reagents/Chemistry-Reagents/Chemistry-Reagents-Food-Drinks.dm
#	code/modules/reagents/reagent_containers/borghydro.dm
#	code/modules/recycling/conveyor2.dm
#	code/modules/recycling/disposal.dm
#	code/modules/research/circuitprinter.dm
#	code/modules/species/species.dm
#	code/modules/species/station/prometheans.dm
#	code/modules/species/xenomorphs/alien_powers.dm
#	icons/atmos/tank.dmi
#	icons/mecha/mech_construct.dmi
#	icons/mecha/mech_construction.dmi
#	icons/mecha/mecha.dmi
#	icons/mob/human_races/species/xenos/eyes.dmi
#	icons/mob/screen/midnight.dmi
#	icons/obj/chemical.dmi
#	icons/obj/clothing/obj_belt.dmi
#	icons/obj/furniture.dmi
#	icons/turf/areas.dmi
#	interface/skin.dmf
#	maps/away/skrellscoutship/skrellscoutship_shuttles.dm
#	maps/torch/torch_define.dm
#	test/check-paths.sh
#	test/run-test.sh
#	tools/dmitool/dmitool.py

+++ code/game/objects/items/weapons/tanks/tanks.dm
@@ -63,7 +63,8 @@ var/list/global/tank_gauge_cache = list()
 	if(istype(loc, /obj/item/device/transfer_valve))
 		var/obj/item/device/transfer_valve/TTV = loc
 		TTV.remove_tank(src)
-		qdel(TTV)
+		if(!QDELETED(TTV)) // It will delete tanks inside it on qdel.
+			qdel(TTV)
 
 	. = ..()
 
@@ -160,7 +161,7 @@ var/list/global/tank_gauge_cache = list()
 			if(do_after(user, 50, src))
 				to_chat(user, "<span class='notice'>You finish attaching the assembly to \the [src].</span>")
 				GLOB.bombers += "[key_name(user)] attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]"
-				message_admins("[key_name_admin(user)] attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]")
+				log_and_message_admins("attached an assembly to a wired [src]. Temp: [air_contents.temperature-T0C]", user)
 				assemble_bomb(W,user)
 			else
 				to_chat(user, "<span class='notice'>You stop attaching the assembly.</span>")
@@ -178,7 +179,7 @@ var/list/global/tank_gauge_cache = list()
 					leaking = 0
 				else
 					GLOB.bombers += "[key_name(user)] attempted to weld a [src]. [air_contents.temperature-T0C]"
-					message_admins("[key_name_admin(user)] attempted to weld a [src]. [air_contents.temperature-T0C]")
+					log_and_message_admins("attempted to weld a [src]. [air_contents.temperature-T0C]", user)
 					if(WT.welding)
 						to_chat(user, "<span class='danger'>You accidentally rake \the [W] across \the [src]!</span>")
 						maxintegrity -= rand(2,6)
@@ -187,6 +188,13 @@ var/list/global/tank_gauge_cache = list()
 			else
 				to_chat(user, "<span class='notice'>The emergency pressure relief valve has already been welded.</span>")
 		add_fingerprint(user)
+
+	if(istype(W, /obj/item/weapon/flamethrower))
+		var/obj/item/weapon/flamethrower/F = W
+		if(!F.status || F.tank || !user.unEquip(src, F))
+			return
+		master = F
+		F.tank = src
 
 /obj/item/weapon/tank/attack_self(mob/user as mob)
 	add_fingerprint(user)
@@ -216,7 +224,7 @@ var/list/global/tank_gauge_cache = list()
 	var/data[0]
 	data["tankPressure"] = round(air_contents && air_contents.return_pressure() ? air_contents.return_pressure() : 0)
 	data["releasePressure"] = round(distribute_pressure ? distribute_pressure : 0)
-	data["defaultReleasePressure"] = round(TANK_DEFAULT_RELEASE_PRESSURE)
+	data["defaultReleasePressure"] = round(initial(distribute_pressure))
 	data["maxReleasePressure"] = round(TANK_MAX_RELEASE_PRESSURE)
 	data["valveOpen"] = using_internal ? 1 : 0
 	data["maskConnected"] = 0
@@ -260,7 +268,7 @@ var/list/global/tank_gauge_cache = list()
 /obj/item/weapon/tank/OnTopic(user, href_list)
 	if (href_list["dist_p"])
 		if (href_list["dist_p"] == "reset")
-			distribute_pressure = TANK_DEFAULT_RELEASE_PRESSURE
+			distribute_pressure = initial(distribute_pressure)
 		else if (href_list["dist_p"] == "max")
 			distribute_pressure = TANK_MAX_RELEASE_PRESSURE
 		else
@@ -274,6 +282,8 @@ var/list/global/tank_gauge_cache = list()
 		return TOPIC_REFRESH
 
 /obj/item/weapon/tank/proc/toggle_valve(var/mob/user)
+
+	var/mob/living/carbon/location
 	if(istype(loc,/mob/living/carbon))
 		var/mob/living/carbon/location = loc
 		if(location.internal == src)
@@ -385,8 +395,7 @@ var/list/global/tank_gauge_cache = list()
 	if(pressure > TANK_FRAGMENT_PRESSURE)
 		if(integrity <= 7)
 			if(!istype(loc,/obj/item/device/transfer_valve))
-				message_admins("Explosive tank rupture! last key to touch the tank was [fingerprintslast].")
-				log_game("Explosive tank rupture! last key to touch the tank was [fingerprintslast].")
+				log_and_message_admins("Explosive tank rupture! last key to touch the tank was [fingerprintslast].")
 
 			//Give the gas a chance to build up more pressure through reacting
 			air_contents.react()
